refactor(web): 响应第七轮审查建议
- 将频道刷新编排逻辑(平台同步 + 失败回退 + 缓存更新 + 通知)提取为 useChatChannelRefresh hook,组件仅保留 UI 状态管理
